additional adjustments
