+++ nym-node/src/node/lp/control/ingress/mod.rs
@@ -1,6 +1,6 @@
 // Copyright 2026 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: GPL-3.0-only
 
 pub mod client_handler;
 pub(crate) mod listener;
-mod node_handler;
+pub mod node_handler;
